url constants extracted
