+++ src/Scraping/index.js
@@ -1,5 +1,6 @@
 const puppeteer = require('puppeteer');
 const fs = require('fs');
+const departamentos = require('../constants/url');
 
 const dir = './screenshots';
 if (!fs.existsSync(dir)) {
@@ -43,37 +44,6 @@ async function getGuaraniesData() {
 
   await page.exposeFunction('getObjetoDeInformacionDeLaMesa', getObjetoDeInformacionDeLaMesa);
   await page.exposeFunction('getInfoCompleta', getInfoCompleta);
-
-  const departamentos = [
-    {
-      name: 'Exactas',
-      url: 'https://g3w.exa.unicen.edu.ar/guarani3w/fecha_examen',
-    },
-    {
-      name: 'Humanas',
-      url: 'http://guaraniweb.fch.unicen.edu.ar/guarani/fecha_examen',
-    },
-    {
-      name: 'Economicas',
-      url: 'https://guarani.econ.unicen.edu.ar/guarani3w/fecha_examen',
-    },
-    {
-      name: 'Veterinarias',
-      url: 'http://serversiu.vet.unicen.edu.ar/guarani/fecha_examen',
-    },
-    {
-      name: 'Ingenieria',
-      url: 'https://guaranitest.fio.unicen.edu.ar/grado/fecha_examen',
-    },
-    {
-      name: 'Derecho',
-      url: 'http://autogestion.der.unicen.edu.ar/guarani/fecha_examen',
-    },
-    {
-      name: 'Salud',
-      url: 'http://guarani.salud.unicen.edu.ar/guarani/fecha_examen',
-    },
-  ];
 
   let informationToWrite = []
