+++ src/constants/url.js
@@ -1,0 +1,34 @@
+const departamentos = [
+  {
+    name: 'Exactas',
+    url: 'https://g3w.exa.unicen.edu.ar/guarani3w/fecha_examen',
+  },
+  {
+    name: 'Humanas',
+    url: 'http://guaraniweb.fch.unicen.edu.ar/guarani/fecha_examen',
+  },
+  {
+    name: 'Economicas',
+    url: 'https://guarani.econ.unicen.edu.ar/guarani3w/fecha_examen',
+  },
+  {
+    name: 'Veterinarias',
+    url: 'http://serversiu.vet.unicen.edu.ar/guarani/fecha_examen',
+  },
+  {
+    name: 'Ingenieria',
+    url: 'https://guaranitest.fio.unicen.edu.ar/grado/fecha_examen',
+  },
+  {
+    name: 'Derecho',
+    url: 'http://autogestion.der.unicen.edu.ar/guarani/fecha_examen',
+  },
+  {
+    name: 'Salud',
+    url: 'http://guarani.salud.unicen.edu.ar/guarani/fecha_examen',
+  },
+];
+
+module.exports = {
+  departamentos,
+}
